+++ w-near/tests/workspaces/main.rs
@@ -1,0 +1,5 @@
+mod test_legacy;
+mod test_storage;
+mod test_transfer;
+mod test_w_near;
+mod utils;
